Refactor project structure and update component imports
- Updated import paths for Header and Updates components in App.jsx
- Deleted unused ButtonLinks, Card, and UpdateCard components along with their styles
- Refactored ContactDetails to use the new ButtonLinks path
- Simplified FooterBackground text
- Removed old styles and components related to buttons and cards
- Created new ButtonLinks component with styles
- Reorganized Navbar component and styles
- Updated AboutMe and Projects sections to reflect new component paths
- Added new styles for updates and refactored Updates section
