+++ src/Components/Buttons/NetlifyButton.jsx
@@ -2,7 +2,12 @@ import "./NetlifyButton.css";
 
 export const NetlifyButton = ({ url }) => {
   return (
-    <a href={url} target="_blank" rel="noopener noreferrer" className="button">
+    <a
+      href={url}
+      target="_blank"
+      rel="noopener noreferrer"
+      className="netifly-button"
+    >
       View Page
     </a>
   );
