+++ src/Components/Buttons/PostButton.css
@@ -1,0 +1,24 @@
+.post-button {
+  border-radius: 1rem;
+  box-shadow: var(--shadow);
+  cursor: pointer;
+  font-family: "IBM Plex Mono", sans-serif;
+  font-size: 1rem;
+  font-weight: 600;
+  padding: 1rem 2rem;
+  text-align: center;
+  background: var(--color-neutral);
+  color: var(--font-color-primary);
+  border: solid 2px var(--color-primary);
+}
+.post-button:hover {
+  background: var(--color-primary);
+  color: var(--font-color-secondary);
+}
+.post-button:active {
+  border: transparent;
+  box-shadow: inset var(--shadow);
+}
+.post-button > li {
+  list-style-type: none;
+}
