+++ src/Components/Buttons/PostButton.jsx
@@ -1,0 +1,14 @@
+import "./PostButton.css";
+
+export const PostButton = ({ url }) => {
+  return (
+    <a
+      href={url}
+      target="_blank"
+      rel="noopener noreferrer"
+      className="post-button"
+    >
+      View Post
+    </a>
+  );
+};
