+++ src/Components/FooterBackground.jsx
@@ -1,8 +1,7 @@
 export const FooterBackground = () => {
   return (
     <div className="footer-background">
-      <p>Lets</p>
-      <p>Connect</p>
+      <p>Lets Connect</p>
     </div>
   );
 };
