+++ src/Data/Updates.json
@@ -1,5 +1,5 @@
 {
-  "Updates": [
+  "updates": [
     {
       "index": 0,
       "name": "Techingo Week 1",
