+++ src/Sections/Footer/FooterContent.css
@@ -5,9 +5,8 @@
   text-align: center;
   padding: 1rem;
   width: auto;
-  z-index: -1;
+  z-index: 1;
   flex-direction: column;
-  background: var(--color-secondry);
 }
 
 .footer-background > p {
@@ -17,25 +16,17 @@
   text-shadow: var(--shadow);
   margin: 0.5rem;
   font-weight: 700;
+  width: 100vw;
 }
-.contact {
+
+.button-links {
+  text-decoration: none;
+}
+.footer-content {
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  justify-content: start;
   align-items: start;
   text-align: start;
   padding: 1rem;
-}
-
-.contact > ul {
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-  padding: 0;
-  margin: 0;
-}
-.contact > ul > li {
-  color: var(--font-color-light);
-  font-size: 1rem;
 }
