+++ src/Sections/Header/Head.jsx
@@ -1,10 +1,9 @@
-import { Navbar } from "../Components/Navbar";
-
+import { Navbar } from "../../Components/NavBar/Navbar";
 
 export const Header = () => {
   return (
     <header className="header">
-      <Navbar/>
+      <Navbar />
     </header>
   );
-}
+};
