+++ src/Sections/MainBody/Updates/Updates.css
@@ -1,4 +1,4 @@
-.updates {
+.updates-container {
   display: flex;
   flex-direction: row;
   justify-content: center;
@@ -6,6 +6,7 @@
   justify-content: space-evenly;
   flex-wrap: wrap;
   padding: 1rem;
+  gap: 1rem;
 }
 .card {
   display: flex;
