+++ src/Sections/MainBody/Updates/Updates.jsx
@@ -1,15 +1,17 @@
-import { UpdateCard } from "../Components/UpdateCard";
+import { UpdateCard } from "../../../Components/Cards/UpdateCard";
 import "./updates.css";
-import updates from "../data/Updates.json";
+import updates from "../../../data/Updates.json";
+
+// Ensure updates has a valid structure
 
 export const Updates = () => {
   return (
-    <section>
+    <section id="updates">
       <div className="h3-title">
         <h2>Updates</h2>
       </div>
       <div className="updates-container">
-        {updates.projects.map((update, index) => (
+        {updates.updates.map((update, index) => (
           <div key={index} className="updates-card">
             <UpdateCard
               icon={update.icon}
