+++ src/Sections/Skills.jsx
@@ -1,7 +1,7 @@
 export const Skills = () => {
   return (
-    <>
+    <section id="skills">
       <h2>Skills</h2>
-    </>
+    </section>
   );
 };
