+++ src/Sections/Toolbox.jsx
@@ -1,7 +1,7 @@
 export const Toolbox = () => {
   return (
-    <>
+    <section id="toolbox">
       <h2> My Toolbox</h2>
-    </>
+    </section>
   );
 };
